Выполнил задание 9.8. Помощь друга
