+++ index.html
@@ -112,8 +112,8 @@
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="type">Тип жилья</label>
           <select id="type" name="type">
-            <option value="bungalow">Бунгало</option>
-            <option value="flat" selected>Квартира</option>
+            <option value="bungalow" selected>Бунгало</option>
+            <option value="flat">Квартира</option>
             <option value="hotel">Отель</option>
             <option value="house">Дом</option>
             <option value="palace">Дворец</option>
@@ -121,7 +121,7 @@
         </fieldset>
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="price">Цена за ночь, руб.</label>
-          <input id="price" name="price" type="number" placeholder="1000" min="0" max="100000" step="100" required
+          <input id="price" name="price" type="number" placeholder="0" min="0" max="100000" step="100" required
             data-pristine-required-message="Вы пропустили это поле(Сообщение Pristin)"
             data-pristine-min-message="Недостаточная сумма(Сообщение Pristin)"
             data-pristine-max-message="Превышено максимальное значение — 100 000(Сообщение Pristin)">
