+++ js/main.js
@@ -1,5 +1,4 @@
 import { disablingAdForm, disablingFormMapFilter } from './form.js';
-import './validation-pristine.js';
 import './map.js';
 import './slider-form.js';
