+++ js/map.js
@@ -14,7 +14,7 @@ const map = L.map('map-canvas')
   .setView({
     lat: 35.6895,
     lng: 139.752465,
-  }, 12);
+  }, 11);
 
 L.tileLayer( //  создаём нужный слой  командой L.tileLayer(), изображениями карт от OpenStreetMap добавив как слой на нашу созданную карту
   'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
@@ -62,8 +62,9 @@ resetButton.addEventListener('click', () => {
   map.setView({ //  возвращение к начальным значениям масштаба и центра карты
     lat: 35.6895,
     lng: 139.692,
-  }, 12);
+  }, 11);
 });
+
 
 // ****************************** Дополнительные пины **********************************
