+++ js/slider-form.js
@@ -2,7 +2,7 @@ const sliderElement = document.querySelector('.ad-form__slider');
 const price = document.querySelector('#price'); // Цена за ночь
 const resetButton = document.querySelector('.ad-form__reset');
 
-price.value = 1000; // Начальное значение в поле ввода
+// price.value = 1000; // Начальное значение в поле ввода
 
 noUiSlider.create(sliderElement, { // Создадим слайдер
   range: {
@@ -22,12 +22,13 @@ noUiSlider.create(sliderElement, { // Создадим слайдер
   },
 });
 
-sliderElement.noUiSlider.on('update', () => { // слушатель события update, которое будет вызвано при изменении положения слайдера
+sliderElement.noUiSlider.on('slide', () => { // слушатель события update(в моём случае slide), которое будет вызвано при изменении положения слайдера
   price.value = sliderElement.noUiSlider.get();
 });
 
 // сброс карты(RESET)
 resetButton.addEventListener('click', (evt) => {
   evt.preventDefault();
+
   sliderElement.noUiSlider.reset(); // начальное значение устанавливать вызовом метода .set() у свойства noUiSlider
 });
